fix imports, rename file name to match class name, remove unused methods from DB2BaseHandler
